# Conflicts:
#   frontend/src/store.js

Former-commit-id: a15bd7fa1cdc03e45467d6a698d5f0dcde8998d0

+++ backend/spring/src/main/java/com/osds/peamo/model/network/request/PerfumeListSearch.java
@@ -9,7 +9,7 @@ import lombok.*;
 public class PerfumeListSearch {
 
     int gender;
-    long category;
+    int category;
     String uid;
 
 }
